+++ cucumber-gherkin/src/main/java/io/cucumber/core/gherkin/Step.java
@@ -2,6 +2,8 @@ package io.cucumber.core.gherkin;
 
 import org.jspecify.annotations.Nullable;
 
+import java.util.List;
+
 public interface Step extends io.cucumber.plugin.event.Step {
 
     StepType getType();
@@ -14,4 +16,5 @@ public interface Step extends io.cucumber.plugin.event.Step {
     @Nullable
     Argument getArgument();
 
+    List<Argument> getArguments();
 }